requirements.txt
glob2
numpy
astropy
pillow
pandas
